+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 6. Add User and Thought Routes
 7. Added Controllers - some not ready & so commented out
 8. Updated User Route & userController to add thought and update user
+9. Updated User Route & userController to add Friend